lab1/requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
flake8==3.8.4
idna==2.9
importlib-metadata==3.4.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.2.0
mccabe==0.6.1
numpy==1.18.1
Pillow==7.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.23.0
scikit-learn==0.24.0
scipy==1.4.1
six==1.14.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.25.8
zipp==3.4.0
